feat(rn-device): platform-split in runAgentDevice — iOS uses rn-fast-runner
* Insert iOS short-circuit at top of runAgentDevice: iOS commands
  (snapshot/tap/press/fill/type/back/screenshot/keyboard) route through
  runIOS() instead of the legacy 3-tier dispatch.
* Move runner-state exports (FastRunnerState, startFastRunner,
  stopFastRunner, getFastRunnerState, parseReadySignal,
  createReadySignalParser, probeFastRunnerLiveness, reapStaleFastRunner,
  fastTap/fastType/fastSwipe/fastSnapshot/fastScreenshot/
  fastDismissKeyboard/fastHealthCheck/isFastRunnerAvailable) from
  fast-runner-session.ts into runners/rn-fast-runner-client.ts.
* Delete src/fast-runner-session.ts + dist/fast-runner-session.js.
* Delete iOS branches from the daemon + CLI dispatch tiers — they
  become dead code with the short-circuit in place. tryFastRunner(),
  cacheRefMapFromResult(), computeSwipeCoords() removed.
* Add STALE_REF error code to ToolErrorCode (returned when a press/
  fill is invoked with a @ref that's no longer in the ref-map).
* Update runIOS snapshot path to map runner SnapshotNodes →
  FlatNode[] and call updateRefMapFromFlat() so press/fill can
  resolve refs from the latest snapshot. (The Swift runner already
  returns flat nodes with index/rect, not a nested tree — so the
  client constructs @e<index> refs directly without flattenXCUITree.)

Per spec §3.1 + §3.6. The agent-device-wrapper.ts iOS branches are now
truly dead — no iOS request hits the external CLI after this PR ships.
Android continues to use the existing 3-tier dispatch.

Non-trivial decisions made during integration:
* fill and type share one buildRunIOSArgs case — the Swift .type
  command focuses an input at x/y AND types in a single call (see
  RnFastRunnerTests+CommandExecution.swift:429-468), so no separate
  tap is needed. Saves one HTTP round-trip per fill.
* longpress/scroll/swipe are intentionally NOT in
  RN_FAST_RUNNER_COMMANDS — they already bypass runAgentDevice via
  fastSwipe()/etc. inside device-interact.ts.
* STALE_REF handling lives inside runIOS() (driven by a _staleRef
  sentinel in RunIOSArgs) rather than in the wrapper, so any caller
  hitting a stale ref gets the same failResult shape + cachedMetadata.
